requirements.txt
fastapi
requests
bcrypt
pybloom_live
pyjwt